Move entity-type definitions into their own file
